small updates to bridge, but no fixes, small cleanup - duplicate
