+++ app/services/send/types.ts
@@ -1,9 +1,23 @@
-import { TokenWithBalance } from 'store/balance'
-import { BN } from 'avalanche'
+import BN from 'bn.js'
 import { BigNumber } from 'ethers'
-import { SignTransactionRequest } from 'services/wallet/types'
+import { TokenWithBalance } from 'store/balance'
+import {SignTransactionRequest} from 'services/wallet/types';
 
-export interface SendState<T = TokenWithBalance> {
+export enum SendEvent {
+  TX_DETAILS = 'SendEvent: TX_DETAILS'
+}
+
+export interface SendError {
+  error: boolean
+  message: string
+}
+
+export const DEFAULT_SEND_HOOK_ERROR: SendError = {
+  error: false,
+  message: ''
+}
+
+export interface SendState<T extends TokenWithBalance = TokenWithBalance> {
   maxAmount?: BN
   amount?: BN
   address?: string
@@ -12,6 +26,7 @@ export interface SendState<T = TokenWithBalance> {
   gasPrice?: BigNumber
   gasLimit?: number
   canSubmit?: boolean
+  loading?: boolean
   token?: T
   txId?: string
 }
@@ -21,14 +36,16 @@ export type ValidSendState = SendState &
     canSubmit: true
   }
 
-export interface SendError {
-  error: boolean
-  message: string
+export function isValidSendState(
+  sendState: SendState
+): sendState is ValidSendState {
+  return sendState.canSubmit === true
 }
 
-export interface SendServiceHelper {
-  getTransactionRequest(sendState: SendState): Promise<SignTransactionRequest>
-  validateStateAndCalculateFees(sendState: SendState): Promise<SendState>
+export interface SendErrors {
+  amountError: SendError
+  addressError: SendError
+  formError: SendError
 }
 
 export enum SendErrorMessage {
@@ -38,4 +55,17 @@ export enum SendErrorMessage {
   INVALID_ADDRESS = 'Address is invalid',
   INVALID_NETWORK_FEE = 'Network Fee is invalid',
   INSUFFICIENT_BALANCE = 'Insufficient balance.'
+}
+
+export interface SendServiceHelper {
+  getTransactionRequest(
+    sendState: SendState,
+    isMainnet: boolean,
+    fromAddress: string
+  ): Promise<SignTransactionRequest>
+  validateStateAndCalculateFees(
+    sendState: SendState,
+    isMainnet: boolean,
+    fromAddress: string
+  ): Promise<SendState>
 }
